refactor(ai/review): split chunker into focused package
The chunker grew to ~1.5k LOC mixing two concerns. Split it into a
`chunker/` package: `grouping.py` (semantic grouping and token-budget
chunking) and `workflow_scripts.py` (workflow script-reference matching
and shell/action-path parsing). Behavior is unchanged; the public API is
preserved via the package `__init__` re-export and the lazy export map
now points at the defining submodule.
