+++ pyproject.toml
@@ -128,6 +128,7 @@ packages = [
   "lintro.ai.prompts",
   "lintro.ai.providers",
   "lintro.ai.review",
+  "lintro.ai.review.chunker",
   "lintro.ai.review.context",
   "lintro.ai.review.enums",
   "lintro.ai.review.models",
